Enable external valdiation of other data
The current proposed Proton 2.x spec supports the addiiton of an
(optional) extra root key 'other' which can contain whatever other
data the user wants. Enable separate validation of the input form
based on this input by reconfiguring the scorer API to being class-
like rather than a single callable. Thankfully this is the shape of
the existing implementations anyway.

Change-Id: I7bb93ef2cf809a6758ed4a8e197c4eaa5005c4f0